+++ includes/features/class-dlogger.php
@@ -325,6 +325,7 @@ class DLogger {
 				'code'        => (int) $code,
 				'environment' => (string) Environment::stage(),
 				'traceID'     => (string) DECALOG_TRACEID,
+				'instance'    => (string) DECALOG_TRACEID,
 			];
 			$channel = $this->current_channel_tag();
 			if ( $this->logger->getName() !== $channel ) {