+++ includes/formatters/class-lokiformatter.php
@@ -91,17 +91,17 @@ class LokiFormatter implements FormatterInterface {
 		switch ( $this->template ) {
 			case 1:
 				$stream['job']      = $this->job;
-				$stream['instance'] = gethostname();
+				$stream['instance'] = DECALOG_INSTANCE_NAME;
 				$stream['level']    = $level_class;
 				break;
 			case 2:
 				$stream['job']      = $this->job;
-				$stream['instance'] = gethostname();
+				$stream['instance'] = DECALOG_INSTANCE_NAME;
 				$stream['env']      = Environment::stage();
 				break;
 			case 3:
 				$stream['job']      = $this->job;
-				$stream['instance'] = gethostname();
+				$stream['instance'] = DECALOG_INSTANCE_NAME;
 				$stream['version']  = Environment::wordpress_version_text( true );
 				break;
 			case 4:
@@ -115,7 +115,7 @@ class LokiFormatter implements FormatterInterface {
 				break;
 			default:
 				$stream['job']      = $this->job;
-				$stream['instance'] = gethostname();
+				$stream['instance'] = DECALOG_INSTANCE_NAME;
 		}
 		$date             = new \DateTime();
 		$values[]         = (string) ( $date->format( 'Uu' ) * 1000 );