+++ includes/handlers/class-grafanamonitoringhandler.php
@@ -68,12 +68,12 @@ class GrafanaMonitoringHandler extends AbstractMonitoringHandler {
 		switch ( $this->template ) {
 			case 1:
 				$stream['job']         = $this->job;
-				$stream['instance']    = gethostname();
+				$stream['instance']    = DECALOG_INSTANCE_NAME;
 				$stream['environment'] = Environment::stage();
 				break;
 			case 2:
 				$stream['job']      = $this->job;
-				$stream['instance'] = gethostname();
+				$stream['instance'] = DECALOG_INSTANCE_NAME;
 				$stream['version']  = Environment::wordpress_version_text( true );
 				break;
 			case 3:
@@ -82,7 +82,7 @@ class GrafanaMonitoringHandler extends AbstractMonitoringHandler {
 				break;
 			default:
 				$stream['job']      = $this->job;
-				$stream['instance'] = gethostname();
+				$stream['instance'] = DECALOG_INSTANCE_NAME;
 		}
 		foreach ( $stream as $key => $value ) {
 			//$this->endpoint .= '/' . $key . '/' . $value;